refactor: extract audio, MIDI and launchpad hooks
